Add processor module, to start processing the queue (not done yet).
The processor uses a QueueProcessor to process data on the queue (filled
with collectors). The processor uses StatsShelf to keep track of

stats, so stats are calculated as they're added to the shelf.
Later the processor will use other backends to flush the stats stored in
the shelf.

From now on we have dependency on statsdmetrics module as well (to parse
        and operate on stats metrics).