Writing tests for queue processor, added loggin to queue processor
Not finished yet. The processor now can be called to shutdown and has
a method to wait for the shutdown (so is used in tests).
However the test does for processing does not pass yet since flush() is
not being called.
But I want to commit the code, so I skipped the test.