Add a module to define the server, and layout how we used it in app
The server is supposed to be the connecting point between collectors and
the queue processor.
I'm still trying to find a good way of managing them, so the idea is not
final yet.
The server will be handling how collectors and queue processor will
interact, how to reload (gracefully) the collectors and the queue
processor and everything.
The app.App class (which is the main application) demonstrates how the
server is supposed to be working.
Also I have this fealing that there could be multiple server classes to
manage in different way. The default server will start the collectors
in a process and the queue processor in another.
But I think a lightweight (threaded) version can also be implemented
(just spawning collectors will be different) to use fewer system
resources if the load is not that much.
Another server (Say MultiProcessCollecor) can be implemented to start
each collector in a different process (so we can have mutliple
        collectors, each one running on a different CPU).