App configures logger, and creates the destinations specs for server
The app configures the server (its logger, destinations, etc) then
runs the server.
Now when the server runs it starts collectors and makes sure that enough
destinations are available to meet the requested configurations.
This is so that the server can be used to reload in a graceful way,
     recreating the destinations, and recreate collectors (if required),
without loosing the queue.