Add a test for QueueProcessor for resuming after shutdown
To make sure we can shutdown the processor, keep inserting items to the
queue, and then if we modify the processor and then run the same processor,
it will continue processing the same queue correctly.
This test is used for reloading the server, where we do not want to
recreate the queue or the processor, but we'd like to change the
destinations of the processor.