Remove destinations from server, add set_destinations to queue processor
There is no need to keep a reference to the destinations in the server
now), so remove the attribute (and from the test).
I think we'd need to be able to set the destinations to the queue
processor at once (instead of just clearing them and adding them one by one)
Since it might cause the processor to flush to old destinations, while
we're adding new destinations to the procesor in another thread.