Fix queue processor stapping by default when queue is empty
Since I'd picked the default stop queue token to be None, and if queue
is empty we used to set the data to None, the first time that the queue
is empty, it's considered as a stop signal and stops the processor.
Fix this by not using None when the queue is empyt.