Solve signal handling issue with the app
As the main thread started the server, and after starting queue and
collectors, the server would join() those threads, this blocks the main
thread to run the signal handlers.

So we'll start the server, in a separate thread from the app.run().
But again if we wait for that thread to join(), same issue happens.

My solution (so far) is to not wait for the server using a thread.join()
block, but instead use a blocking syscall (that will be interrupted
by the OS when a signal is sent). So we poll if the server thread is
still running and block with sleep() instead.

This way, while the main thread is in sleep(), if a signal is sent, OS
will wake the thread and call the signal handlers as expected.