Add a TaskThreadPool class to system utils (on going)
To help run some tasks concurrently easier by prestarting some worker
threads. This is going to be used in TCP server so we can pre start some
threads and handle TCP requests in them.

The task is not done yet, the ThreadPool will also have to spawn new
threads if queue is not empty and we need more tests.