Resolve ResourceWarning in Python 3, for UDP socket collector shutdown
Since the UDP server socket is not actually "connected", calling
shutdown() on it raised an exception and it caused the call to
socket.close() get skipped silently. So Python would close the socket
automatically at exit but showing a ResourceWarning about it.
Solved it by making sure socket.close() is called on the socket server
regardless if calling shutdown on the socket was successful or not.

Also improve a bit on the local references in the UDP server collection
loop.