TCP collectors handle incomplete lines by buffering the request
Before this commit, TCP collectors were acting like UDP collectors,
trying to read as much as possible from the connection, and put all the
available data into the process queue.

This might cause an issue where a request from the client does not come
in 1 go (client is sending metrics with pauses in between sends) and the
request might not contain full lines of metrics. There might be
possibility of losing incompelte lines.

This commit tries to read from TCP connection in smaller chunks (8KB)
and if the request is not finished with a complete line, it assumes a
later send might provide the rest of the line.

Finally, the last incomplete line is also enqueued to be processed,
implying that the last metric from the client does not need to end
with a new line.