Refactor server start/shutdown collectors/processor to accept params
So that the same methods to start/shutdown collectors or queue processor
can be used to start/shutdown other objects, not just the ones
associated with the server.

This is to be used in implementing reload() method of the server, where
we'll have to manage collectors and queue processors, but not only the
ones set for the server, but also the ones that used to be associated to
the server, but not anymore.