Implementing a functional test for reloading the server (WIP)
The test describes the expected behavior, when we send SIGHUP to the
server process the server should reload by rereading the configuration
and stopping/starting collectors and queue processor.

This still does not work since the queue processor itself does not
support reloading (by resetting destinatoins), so when the app creates
new destinations and sets it the queue processor so the server can be
called to reload(), it failes as the state of the queue processor is
running.

We need to fix it either by setting the destinations on the server
itself, or add support for setting new destinations to the queue
processor while the processor is still running.