Allow setting destinations for the queue processor while it's running
Before calling set_destinations() on the queue processor while it was running
used to throw exceptions. Now it'll stop the flush threads, and sets  the destinations
and then initializes the destinations once again.

This is useful for reloading the server, where the app needs to change the destinations
for the queue processor while reloading.