Fixed queue processor failing to reload correctly
I forgot to clear the event that indicated the process method should
stop, after processing was finished. So when the queue processor
stopped, next calls on the process method were not actually working.
We had a test for this case, but the test was not asserting flush
destinations, so it could not detect this flaw.